deepseek_rag_project/backend/requirements.txt
# 基础框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
pydantic>=2.6.0
python-dotenv
aiofiles

# 关键修复：锁定 Numpy < 2.0 防止冲突
numpy<2.0.0
tenacity<9.0.0

# LlamaIndex 核心组件
llama-index==0.10.55
llama-index-core==0.10.55
llama-index-embeddings-huggingface==0.2.2
llama-index-vector-stores-milvus==0.1.10
llama-index-readers-file==0.1.25

# 🚀 新增：本地模型支持 (Ollama)
llama-index-llms-ollama

# 🚀 新增：OCR 与 视频处理
rapidocr_onnxruntime
opencv-python-headless
moviepy
faster-whisper
timm
einops
accelerate

# 基础工具
pymilvus>=2.3.0
torch>=2.2.0
transformers
